Separated dashboards and dashboard from pydash.web.controller.dashboards.py
