+++ pydash/pydash_web/controller/__init__.py
@@ -12,7 +12,7 @@ from .change_settings import change_settings
 from .change_password import change_password
 
 from .dashboards import dashboards
-from .dashboards import dashboard
+from .dashboard import dashboard
 from .register_dashboard import register_dashboard
 from .change_dashboard_settings import change_dashboard_settings
 from .delete_dashboard import delete_dashboard
